catalog: carve VisualizationService out of the facade
Move catalog visualizations + dashboards (CRUD, enrichment), worker
compute (compute_saved_visualization_rows, compute_ad_hoc_visualization,
get_visualization_fields, get_table_column_stats), source descriptor
resolution (_resolve_source_for_worker, _build_sql_worker_source) and
the GW raw-select fallback into catalog/services/visualizations.py.

Apply Category C bare-except cleanup in _build_sql_worker_source via
the resolve_or_log helper.

Worker triggers (trigger_visualize_query, trigger_visualize_fields,
trigger_visualize_column_stats, trigger_resolve_virtual_table) are
looked up lazily off the service module so test monkeypatches keep
working. VisualizationService routes resolve_virtual_flow_table
through a facade reference (bound after construction) so tests that
patch CatalogService.resolve_virtual_flow_table at the class level
still take effect.
